Traktorentest
git-svn-id: https://subversion.assembla.com/svn/oop-ws2012@250 8f0689cb-c7f0-493e-b951-6e7372201282

+++ OOP8/src/BiogasTraktor.java
@@ -5,12 +5,12 @@ public class BiogasTraktor extends Traktor {
 		super(nummer, einsatzzweck);
 	}
 
-	public Float getBisherigerVerbrauch() {
-		return bisherigerVerbrauch.floatValue();
+	public double getBisherigerVerbrauch() {
+		return bisherigerVerbrauch.doubleValue();
 	}
 
-	public void erhoeheVerbrauch(float biogas) {
-		bisherigerVerbrauch = bisherigerVerbrauch.floatValue() + biogas;
+	public void erhoeheVerbrauch(double biogas) {
+		bisherigerVerbrauch = bisherigerVerbrauch.doubleValue() + biogas;
 	}
 
 }
